Create README.md
Hanuman, in Hindu mythology, the monkey commander of the monkey army. His exploits are narrated in the great Hindu Sanskrit poem the Ramayana (“Rama's Journey”). While still a baby, Hanuman, the child of a nymph by the wind god, tried to fly up and grab the Sun, which he mistook for a fruit.